Overall, finished the basics to search for and add new nodes to the display. Added up/down buttons for cycling through recipes in a particular process type. Added button to add nodes from the search GUI into the display. Implemented a grid in the background of the display, as well as some help text. Fixed totalEU calculations in Gregtech recipes from going over the max value for integers. Changed how Programmed Circuits are displayed, to include their config number. Fixed bug in Recipe Manager that didn't properly load Programmed Circuits.

+++ core/src/main/java/com/bigbass/nep/recipes/GregtechRecipe.java
@@ -86,8 +86,8 @@ public class GregtechRecipe implements IRecipe {
 		return false;
 	}
 	
-	public int getTotalEU(){
-		return eut * duration;
+	public long getTotalEU(){
+		return ((long) eut) * ((long) duration);
 	}
 	
 	public int getTotalEU(Tier overclock){
